+++ src/main/java/konkuk/thip/common/security/resolver/UserIdArgumentResolver.java
@@ -1,10 +1,9 @@
-package konkuk.thip.common.security.argument_resolver;
+package konkuk.thip.common.security.resolver;
 
 import jakarta.servlet.http.HttpServletRequest;
 import konkuk.thip.common.exception.AuthException;
-import konkuk.thip.common.security.annotation.AuthToken;
+import konkuk.thip.common.security.annotation.UserId;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -15,26 +14,25 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 import static konkuk.thip.common.exception.code.ErrorCode.AUTH_TOKEN_NOT_FOUND;
 
 @Component
-@Slf4j
 @RequiredArgsConstructor
-public class AuthTokenArgumentResolver implements HandlerMethodArgumentResolver {
+public class UserIdArgumentResolver implements HandlerMethodArgumentResolver {
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(AuthToken.class)
-            && parameter.getParameterType().equals(String.class);
+        return parameter.hasParameterAnnotation(UserId.class)
+                && parameter.getParameterType().equals(Long.class);
     }
 
     @Override
-    public String resolveArgument(MethodParameter parameter,
+    public Long resolveArgument(MethodParameter parameter,
                                   ModelAndViewContainer mavContainer,
                                   NativeWebRequest webRequest,
                                   WebDataBinderFactory binderFactory) {
 
-        Object token = ((HttpServletRequest) webRequest.getNativeRequest()).getAttribute("token");
-        if (token == null) {
+        Object userId = ((HttpServletRequest) webRequest.getNativeRequest()).getAttribute("userId");
+        if (userId == null) {
             throw new AuthException(AUTH_TOKEN_NOT_FOUND);
         }
-        return (String) token;
+        return (Long) userId;
     }
 }
